Fix warning
<==== Mercurial data ====>
Branch: meta_4.0

+++ vms/server/plugins/analytics/stub_analytics_plugin/src/nx/vms_server_plugins/analytics/stub/device_agent.cpp
@@ -390,7 +390,7 @@ IMetadataPacket* DeviceAgent::cookSomeEvents()
     auto nextEventTypeIndex =
         [this]()
         {
-            return (m_eventContext.currentEventTypeIndex == kEventsToFire.size() - 1)
+            return (m_eventContext.currentEventTypeIndex == (int) kEventsToFire.size() - 1)
                 ? 0
                 : (m_eventContext.currentEventTypeIndex + 1);
         };
